+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/Knowledgeable.java
@@ -2,12 +2,12 @@ package insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantm
 
 import insane96mcp.iguanatweaksexpanded.data.generator.ISEItemTagsProvider;
 import insane96mcp.iguanatweaksexpanded.module.experience.enchantments.NewEnchantmentsFeature;
-import insane96mcp.iguanatweaksreborn.module.experience.enchantments.enchantment.damage.BonusDamageEnchantment;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.ExperienceOrb;
 import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.enchantment.Enchantment;
@@ -27,7 +27,7 @@ public class Knowledgeable extends Enchantment {
     }
 
     private static float getChance() {
-        return 0.2f;
+        return 0.35f;
     }
 
     @Override
@@ -37,13 +37,13 @@ public class Knowledgeable extends Enchantment {
                 || player.getAttackStrengthScale(0.5f) < 0.9f)
             return;
 
-        float chance = getChance() * BonusDamageEnchantment.getDamageBonusRatio(attacker.getMainHandItem());
+        float chance = (float) (getChance() * (1f / attacker.getAttributeValue(Attributes.ATTACK_SPEED)));
         if (target.getRandom().nextFloat() <= chance)
             target.level().addFreshEntity(new ExperienceOrb(target.level(), target.getX(), target.getY() + target.getBbHeight() / 2f, target.getZ(), lvl));
     }
 
     public static int applyToBlockDrops(Player player, int expToDrop) {
-        float chance = getChance() * BonusDamageEnchantment.getDamageBonusRatio(player.getMainHandItem());
+        float chance = (float) (getChance() * (1f / player.getAttributeValue(Attributes.ATTACK_SPEED)));
         int lvl = EnchantmentHelper.getEnchantmentLevel(NewEnchantmentsFeature.KNOWLEDGEABLE.get(), player);
         return lvl > 0 && player.getRandom().nextFloat() <= chance
                 ? expToDrop + lvl